+++ lib/features/family/features/profiles/widgets/give_bottomsheet.dart
@@ -10,7 +10,7 @@ import 'package:givt_app/features/family/features/giving_flow/organisation_detai
 import 'package:givt_app/features/family/features/impact_groups/cubit/impact_groups_cubit.dart';
 import 'package:givt_app/features/family/features/profiles/cubit/profiles_cubit.dart';
 import 'package:givt_app/features/family/features/profiles/widgets/action_tile.dart';
-import 'package:givt_app/features/family/features/profiles/widgets/empty_wallet_bottom_sheet.dart';
+import 'package:givt_app/features/family/features/topup/screens/empty_wallet_bottom_sheet.dart';
 import 'package:givt_app/utils/utils.dart';
 import 'package:go_router/go_router.dart';
